requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-gis==0.17
fer==21.0.5
idna==3.3
keras==2.6.0
kiwisolver==1.3.2
matplotlib==3.4.3
mtcnn==0.1.1
numpy==1.21.3
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
pandas==1.3.4
Pillow==8.4.0
pkg_resources==0.0.0
postgis==1.0.4
psycopg2==2.9.1
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
urllib3==1.26.7
